# core/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_alpkeskin_rota_core_internal_models.ChartDataPoint:
    properties:
      time:
        type: string
      value:
        type: integer
    type: object
...
  github_com_alpkeskin_rota_core_internal_models.DashboardStats:
    properties:
      active_proxies:
        type: integer
      avg_response_time:
        type: integer
      avg_success_rate:
        type: number
      request_growth:
        type: number
      response_time_delta:
        type: integer
      success_rate_growth:
        type: number
      total_proxies:
        type: integer
      total_requests:
        type: integer
    type: object
  github_com_alpkeskin_rota_core_internal_models.ErrorResponse:
    properties:
      details: {}
      error:
        type: string
    type: object
...
  github_com_alpkeskin_rota_core_internal_models.Log:
    properties:
      details:
        type: string
      id:
        type: integer
      level:
        type: string
      message:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      timestamp:
        type: string
    type: object
  github_com_alpkeskin_rota_core_internal_models.LogListResponse:
    properties:
      logs:
        items:
          $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.Log'
        type: array
      pagination:
        $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.PaginationMeta'
    type: object
...
  github_com_alpkeskin_rota_core_internal_models.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  github_com_alpkeskin_rota_core_internal_models.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.UserInfoResponse'
    type: object
  github_com_alpkeskin_rota_core_internal_models.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  github_com_alpkeskin_rota_core_internal_models.ResponseTimeChartData:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ChartDataPoint'
        type: array
    type: object
...
  github_com_alpkeskin_rota_core_internal_models.SuccessRateChartData:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.SuccessRateDataPoint'
        type: array
    type: object
  github_com_alpkeskin_rota_core_internal_models.SuccessRateDataPoint:
    properties:
      failure:
        type: integer
      success:
        type: integer
      time:
        type: string
    type: object
...
  github_com_alpkeskin_rota_core_internal_models.UserInfoResponse:
    properties:
      username:
        type: string
    type: object
...
host: localhost:8001
info:
  contact:
    name: API Support
    url: https://github.com/alpkeskin/rota
  description: |-
    A high-performance proxy rotation server with health monitoring and intelligent routing
    Provides comprehensive API for managing proxy servers, monitoring their health,
    and configuring rotation strategies.
  license:
    name: LICENSE
    url: https://github.com/alpkeskin/rota/blob/main/LICENSE
  title: Rota Proxy API
  version: 1.0.0
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and receive JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.LoginResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
      summary: User login
      tags:
      - auth
  /dashboard/charts/response-time:
    get:
      description: Get response time chart data for visualization
      parameters:
      - default: 4h
        description: Time interval (e.g., 4h, 24h)
        in: query
        name: interval
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Chart data
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ResponseTimeChartData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
      summary: Response time chart
      tags:
      - dashboard
  /dashboard/charts/success-rate:
    get:
      description: Get success rate chart data for visualization
      parameters:
      - default: 4h
        description: Time interval (e.g., 4h, 24h)
        in: query
        name: interval
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Chart data
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.SuccessRateChartData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
      summary: Success rate chart
      tags:
      - dashboard
  /dashboard/stats:
    get:
      description: Get dashboard statistics including proxy and request metrics
      produces:
      - application/json
      responses:
        "200":
          description: Dashboard statistics
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.DashboardStats'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
      summary: Dashboard statistics
      tags:
      - dashboard
...
  /logs:
    get:
      description: Get paginated list of system logs with optional filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 100
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by log level
        in: query
        name: level
        type: string
      - description: Search term
        in: query
        name: search
        type: string
      - description: Filter by source
        in: query
        name: source
        type: string
      - description: Start time (RFC3339)
        in: query
        name: start_time
        type: string
      - description: End time (RFC3339)
        in: query
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of logs
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.LogListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
      summary: List logs
      tags:
      - logs
  /logs/export:
    get:
      description: Export system logs in various formats (txt, json)
      parameters:
      - default: txt
        description: Export format (txt/json)
        in: query
        name: format
        type: string
      - description: Filter by log level
        in: query
        name: level
        type: string
      - description: Filter by source
        in: query
        name: source
        type: string
      - description: Start time (RFC3339)
        in: query
        name: start_time
        type: string
      - description: End time (RFC3339)
        in: query
        name: end_time
        type: string
      produces:
      - text/plain
      - application/json
      responses:
        "200":
          description: Exported file
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
      summary: Export logs
      tags:
      - logs
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"